pyproject.toml
[tool.poetry]
name = "chumakova_bot"
version = "0.1.0"
description = ""
repository = "some_git_url"
authors = ["Anastasia Chumakova <pospelowa.anast@yandex.ru>"]
packages = [
    {include = "src"},
]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.4"
pydantic = {extras = ["dotenv"], version = "^2.2.0"}
pydantic-settings = "^2.0.3"
loguru = "^0.7.0"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[messages]
hello_pattern = """
Здравствуйте, {user_first_name}!
Рада знакомству!
Для того, чтобы узнать, что умеет делать этот бот, нажмите на кнопку
<u><b><i>Помощь</i></b></u>
"""

help_pattern = """
<u><b>Бот умеет выполнять 9 команд:</b></u>
   1. /start - начало работы с ботом 🤖

   2. Отправьте текст <u><b><i>'фото из старшей школы'</i></b></u> - бот отправит
мою фотографию из старшей школы 👩🏻‍🏫

   3. Отправьте текст <u><b><i>твое последнее селфи</i></b></u> - бот отправит
мое последнее селфи 🤳🏻

   4. Нажмите на кнопку <u><b><i>Что такое GPT</i></b></u> и мой бот вам расскажет
что такое GPT в голосовом сообщении 👾

   5. Нажмите на кнопку <u><b><i>Разница между SQL и NoSQL</i></b></u> и мой бот
вам расскажет разницу между SQL и NoSQL в голосовом сообщении 💾

   6. Нажмите на кнопку <u><b><i>История первой любви</i></b></u> и мой бот
вам расскажет историю моей первой любви в голосовом сообщении ❤️

   7. Нажмите на кнопку <u><b><i>О моем увлечении</i></b></u> и мой бот
вам расскажет о моем увлечении в виде текста 🎨

   8. Нажмите на кнопку <u><b><i>Ссылка на проект</i></b></u> и мой бот
вам пришлет ссылку на проект 💻

   9. Нажмите на кнопку <u><b><i>Помощь</i></b></u> и мой бот
вам пришлет инструкцию по использованию бота 📝

"""

post_url = """<a href="https://telegra.ph/Moe-hobbi-08-17">О моем хобби</a>"""